requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
feedparser==6.0.10
frozenlist==1.3.3
gptcache==0.1.21
grpcio==1.54.0
html2text==2020.1.16
idna==3.4
langchain==0.0.142
llama-index==0.5.27
marshmallow==3.19.0
marshmallow-enum==1.5.1
momento==1.1.1
momento-wire-types==0.39.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
protobuf==4.22.3
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
sgmllib3k==1.0.0
six==1.16.0
SQLAlchemy==1.4.47
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2
